+++ src/components/spartan/index.ts
@@ -42,3 +42,4 @@ export * from './SSkeleton';
 export * from './SInputMask';
 export * from './SInputMaskBlock';
 export * from './SInputDate';
+export * from './SInputDateBlock';